PB - MF-273 - Move phone number tests (still dependent on patient registration)
